+++ processor/src/main/kotlin/com/infinum/collar/processor/extensions/Element.kt
@@ -10,7 +10,11 @@ import javax.lang.model.element.NestingKind
 import javax.lang.model.element.PackageElement
 import javax.lang.model.element.TypeElement
 import javax.lang.model.element.VariableElement
+import javax.lang.model.type.DeclaredType
+import javax.lang.model.type.TypeKind
 import javax.lang.model.util.ElementFilter
+import javax.lang.model.util.Elements
+import javax.lang.model.util.Types
 
 internal fun Element.isSealedClass(): Boolean =
     this.getAnnotation(Metadata::class.java)
@@ -31,22 +35,42 @@ internal fun Element.constructorParameterNames(): List<String> =
 internal fun Element.fieldElements(): List<VariableElement> =
     ElementFilter.fieldsIn(this.enclosedElements).orEmpty()
 
-internal fun VariableElement.resolveMethod(): String =
-    this.asType().toString().let {
-        when (it) {
-            "java.lang.String" -> "putString"
-            "boolean" -> "putBoolean"
-            "byte" -> "putByte"
-            "char" -> "putChar"
-            "double" -> "putDouble"
-            "float" -> "putFloat"
-            "int" -> "putInt"
-            "long" -> "putLong"
-            "short" -> "putShort"
-            "android.os.Bundle" -> "putBundle"
-            else -> ""
+@Suppress("NestedBlockDepth", "ReturnCount")
+internal fun VariableElement.isSupported(typeUtils: Types, elementUtils: Elements): Boolean {
+    val type = this.asType()
+    val kind = this.asType().kind
+
+    if (kind.isPrimitive) {
+        return true
+    } else {
+        if (kind == TypeKind.DECLARED) {
+            val isString = typeUtils.isAssignable(
+                type,
+                elementUtils.getTypeElement(String::class.java.canonicalName).asType()
+            )
+            if (isString) {
+                return true
+            } else {
+                val mapElement: TypeElement = elementUtils.getTypeElement(Map::class.java.canonicalName)
+                val isMap = typeUtils.isAssignable(typeUtils.erasure(type), mapElement.asType())
+                return if (isMap) {
+                    val hasStringKeys = (type as? DeclaredType)?.typeArguments?.firstOrNull()?.let {
+                        typeUtils.isAssignable(
+                            it,
+                            elementUtils.getTypeElement(String::class.java.canonicalName).asType()
+                        )
+                    } ?: false
+
+                    hasStringKeys
+                } else {
+                    false
+                }
+            }
+        } else {
+            return false
         }
     }
+}
 
 internal fun TypeElement.asClassName(): ClassName {
     fun isClassOrInterface(e: Element) = e.kind.isClass || e.kind.isInterface