+++ processor/src/main/kotlin/com/infinum/collar/processor/models/EventParameterHolder.kt
@@ -2,7 +2,7 @@ package com.infinum.collar.processor.models
 
 internal data class EventParameterHolder(
     val enabled: Boolean,
-    val method: String,
+    val isSupported: Boolean,
     val resolvedName: String,
     val variableName: String
 ) : Holder